+++ src/domains/cartography.ts
@@ -61,7 +61,7 @@ export default class CartographyApi {
    *
    * @param {Object} params - The parameters for the request.
    * @param {string} [params.view="compact"] - The view of the buildings.
-   * @returns {Promise<readonly BuildingListElement[]>} - A promise that resolves to a list of building list elements.
+   * @returns Promise<readonly BuildingListElement[]> - A promise that resolves to a list of building list elements.
    */
   getBuildings(
     params: {
@@ -90,7 +90,7 @@ export default class CartographyApi {
    * @param {ID} buildingId - The ID of the building to fetch.
    * @param {Object} [params] - The parameters for the request.
    * @param {string} [params.view="compact"] - The view of the building.
-   * @returns {Promise<Building>} - A promise that resolves to the building information.
+   * @returns Promise<Building> - A promise that resolves to the building information.
    */
   getBuildingById(
     buildingId: ID,
@@ -111,7 +111,7 @@ export default class CartographyApi {
   /**
    * Retrieves the current organization from the API.
    *
-   * @return {Promise<Organization>} A promise that resolves to the current organization.
+   * @returns Promise<Organization> A promise that resolves to the current organization.
    */
   getCurrentOrganization(): Promise<Organization> {
     return this.apiBase
@@ -128,7 +128,7 @@ export default class CartographyApi {
    *
    * @param {ID} buildingId - the building id to update
    * @param {Partial<BuildingForm>} buildingForm - the building information to use
-   * @return {Promise<BuildingListElement>} A promise that resolves to the updated building
+   * @returns Promise<BuildingListElement> A promise that resolves to the updated building
    */
   patchBuilding(
     buildingId: ID,
@@ -145,7 +145,7 @@ export default class CartographyApi {
    * Creates a new building in the API.
    *
    * @param {BuildingForm} buildingForm - The building information used to create it.
-   * @return {Promise<BuildingListElement>} A promise that resolves to the created building.
+   * @returns Promise<BuildingListElement> A promise that resolves to the created building.
    */
   createBuilding(buildingForm: BuildingForm): Promise<BuildingListElement> {
     return this.apiBase
@@ -160,7 +160,7 @@ export default class CartographyApi {
    * Deletes a building given its id
    *
    * @param {ID} buildingId - The building id to delete
-   * @return {Promise<void>} A promise that resolves after deleting the building
+   * @returns Promise<void> A promise that resolves after deleting the building
    */
   deleteBuilding(buildingId: ID): Promise<void> {
     return this.apiBase.delete({
@@ -172,7 +172,7 @@ export default class CartographyApi {
    * Returns a list of floors given a search criteria
    *
    * @param {FloorSearch} params - the list of parameters to filter the response
-   * @return {Promise<readonly Floor[]>} A promise that resolves to a list of floors
+   * @returns Promise<readonly Floor[]> A promise that resolves to a list of floors
    */
   getFloors(params: FloorSearch = {}): Promise<readonly Floor[]> {
     const url =
@@ -191,7 +191,7 @@ export default class CartographyApi {
    * Retrieves a floor by its ID.
    *
    * @param {ID} floorId - The ID of the floor to retrieve
-   * @return {Promise<Floor>} A promise that resolves to the retrieved floor
+   * @returns Promise<Floor> A promise that resolves to the retrieved floor
    */
   getFloorById(floorId: ID): Promise<Floor> {
     return this.apiBase
@@ -206,7 +206,7 @@ export default class CartographyApi {
    *
    * @param {ID} floorId - the floor id to update
    * @param {Partial<FloorForm>} floorForm - the floor information to use
-   * @returns {Promise<Floor>} Promise that resolves to the updated floor
+   * @returns Promise<Floor> Promise that resolves to the updated floor
    */
   patchFloor(floorId: ID, floorForm: FloorForm): Promise<Floor> {
     return this.apiBase
@@ -221,7 +221,7 @@ export default class CartographyApi {
    * Creates a new floor
    *
    * @param {FloorForm} floorForm - The floor information used to create it
-   * @return {Promise<Floor>} Promise that resolves to the created floor
+   * @returns Promise<Floor> Promise that resolves to the created floor
    */
   createFloor(floorForm: FloorForm): Promise<Floor> {
     return this.apiBase
@@ -236,7 +236,7 @@ export default class CartographyApi {
    * Deletes a floor given its ID.
    *
    * @param {ID} floorId - The ID of the floor to delete
-   * @return {Promise<void>} Promise that resolves once the floor is deleted
+   * @returns Promise<void> Promise that resolves once the floor is deleted
    */
   deleteFloor(floorId: ID) {
     return this.apiBase.delete({ url: "/api/v1/floors/" + floorId });
@@ -247,13 +247,14 @@ export default class CartographyApi {
    * If no organizationId is provided in the search parameters, it retrieves the organizationId from the JWT.
    *
    * @param {GeofenceSearch} [params={}] - The search parameters for the geofences.
-   * @returns {Promise<Paginated<Geofence>>} A promise that resolves to a paginated list of geofences.
+   * @returns Promise<Paginated<Geofence>> A promise that resolves to a paginated list of geofences.
    */
   async getGeofences(
     params: GeofenceSearch = {},
   ): Promise<Paginated<Geofence>> {
+    const authSession = await this.apiBase.getAuthSession();
     if (!params.organizationId) {
-      params.organizationId = await this.apiBase.getJwtOrganizationId();
+      params.organizationId = authSession.organizationId;
     }
 
     return this.apiBase
@@ -273,7 +274,7 @@ export default class CartographyApi {
    * Retrieves a geofence by its ID.
    *
    * @param {UUID} geofenceId - The ID of the geofence to retrieve.
-   * @return {Promise<Geofence>} A Promise that resolves to the retrieved geofence.
+   * @returns Promise<Geofence> - A Promise that resolves to the retrieved geofence.
    */
   getGeofenceById(geofenceId: UUID): Promise<Geofence> {
     return this.apiBase
@@ -288,7 +289,7 @@ export default class CartographyApi {
    *
    * @param {UUID} geofenceId - The ID of the geofence to update.
    * @param {Partial<GeofenceForm>} geofenceForm - The partial data to update the geofence with.
-   * @return {Promise<Geofence>} A Promise that resolves to the updated geofence.
+   * @returns Promise<Geofence> - A Promise that resolves to the updated geofence.
    */
   patchGeofence(
     geofenceId: UUID,
@@ -306,7 +307,7 @@ export default class CartographyApi {
    * Creates a new geofence.
    *
    * @param {GeofenceForm} geofenceForm - The geofence information used to create it.
-   * @return {Promise<Geofence>} A Promise that resolves to the created geofence.
+   * @returns Promise<Geofence> A Promise that resolves to the created geofence.
    */
   createGeofence(geofenceForm: GeofenceForm): Promise<Geofence> {
     return this.apiBase
@@ -321,7 +322,7 @@ export default class CartographyApi {
    * Deletes a geofence by its ID.
    *
    * @param {UUID} geofenceId - The ID of the geofence to delete.
-   * @return {Promise<void>}
+   * @returns Promise<void>
    */
   deleteGeofence(geofenceId: UUID) {
     return this.apiBase.delete({ url: "/api/v1/geofences/" + geofenceId });
@@ -331,7 +332,7 @@ export default class CartographyApi {
    * Retrieves paths based on the provided parameters.
    *
    * @param {PathSearch} params - The search parameters for fetching paths.
-   * @return {Promise<Paths[]>} A Promise that resolves to an array of paths.
+   * @returns Promise<Paths[]> - A Promise that resolves to an array of paths.
    */
   getPaths(params: PathSearch = {}): Promise<Paths[]> {
     const url = params.buildingId
@@ -346,7 +347,7 @@ export default class CartographyApi {
    *
    * @param {number} buildingId - the building id to update paths from
    * @param {Paths} pathForm - the path information to use
-   * @returns {Promise<Paths>}
+   * @returns Promise<Paths>
    */
   patchPath(buildingId: number, pathForm: Paths): Promise<Paths> {
     return this.apiBase.put<Paths>({
@@ -359,7 +360,7 @@ export default class CartographyApi {
    * Retrieves a list of Poi objects based on the provided search parameters.
    *
    * @param {PoiSearch} [params] - Optional search parameters to filter the Poi objects.
-   * @returns {Promise<Poi[]>} - A Promise that resolves to an array of Poi objects.
+   * @returns Promise<Poi[]> - A Promise that resolves to an array of Poi objects.
    */
   async getPois(params: PoiSearch = {}): Promise<Poi[]> {
     const url =
@@ -385,7 +386,7 @@ export default class CartographyApi {
    *
    * @param {number} poiId - The ID of the POI to update.
    * @param {PoiUpdateForm} poiForm - The form data containing the updated POI information.
-   * @return {Promise<Poi>} A Promise that resolves to the updated POI object.
+   * @returns Promise<Poi> A Promise that resolves to the updated POI object.
    */
   patchPoi(poiId: number, poiForm: PoiUpdateForm): Promise<Poi> {
     return this.apiBase
@@ -400,7 +401,7 @@ export default class CartographyApi {
    * Creates a new POI
    *
    * @param {PoiCreateForm} poiForm - The poi information used to create it
-   * @return {Promise<Poi>} A Promise that resolves to the created POI object.
+   * @returns Promise<Poi> A Promise that resolves to the created POI object.
    */
   createPoi(poiForm: PoiCreateForm): Promise<Poi> {
     return this.apiBase
@@ -415,7 +416,7 @@ export default class CartographyApi {
    * Deletes a POI given its id.
    *
    * @param {number} poiId - The POI id to delete.
-   * @return {Promise<void>} A Promise that resolves to void.
+   * @returns Promise<void> A Promise that resolves to void.
    */
   deletePoi(poiId: number): Promise<void> {
     return this.apiBase.delete({ url: "/api/v1/pois/" + poiId });
@@ -425,7 +426,7 @@ export default class CartographyApi {
    * Creates an array of POIs in bulk.
    *
    * @param {PoiCreateForm[]} pois - An array of POI objects to create.
-   * @return {Promise<Poi[]>} A Promise that resolves to an array of created POIs.
+   * @returns Promise<Poi[]> A Promise that resolves to an array of created POIs.
    */
   createPoisBulk(pois: PoiCreateForm[]): Promise<Poi[]> {
     return this.apiBase
@@ -440,7 +441,7 @@ export default class CartographyApi {
    * Updates an array of POIs in bulk.
    *
    * @param {PoiUpdateForm[]} poiUpdateForms - An array of POI update forms.
-   * @return {Promise<Poi[]>} An array of Promises that resolves to an array of
+   * @returns Promise<Poi[]> An array of Promises that resolves to an array of
    *                          updated POIs or reject if the update has failed.
    */
   patchPoisBulk(
@@ -461,7 +462,7 @@ export default class CartographyApi {
    * Deletes an array of POIs in batch.
    *
    * @param {number[]} poiIds - An array of POI ids to delete.
-   * @return {Promise<void[]>} A Promise that resolves to void.
+   * @returns Promise<void[]> A Promise that resolves to void.
    */
   deletePoisBulk(poiIds: number[]): Promise<void[]> {
     const deletions = poiIds.map((poiId) =>
@@ -474,7 +475,7 @@ export default class CartographyApi {
   /**
    * Retrieves a list of POI categories.
    *
-   * @return {Promise<PoiCategory[]>} A Promise that resolves to an array of POICategory objects.
+   * @returns Promise<PoiCategory[]> A Promise that resolves to an array of POICategory objects.
    */
   getPoiCategories(): Promise<PoiCategory[]> {
     return this.apiBase
@@ -493,7 +494,7 @@ export default class CartographyApi {
    *
    * @param {number} poiCategoryId - The ID of the POI category to patch.
    * @param {Partial<PoiCategoryForm>} poiCategoryForm - The partial POI category form.
-   * @return {Promise<PoiCategory>} A promise that resolves to the updated POI category.
+   * @returns Promise<PoiCategory> A promise that resolves to the updated POI category.
    */
   patchPoiCategory(
     poiCategoryId: number,
@@ -509,7 +510,7 @@ export default class CartographyApi {
    * Creates a new POI category.
    *
    * @param {PoiCategoryForm} poiCategoryForm - The form data for the POI category to be created.
-   * @return {Promise<PoiCategory>} A promise that resolves to the created POI category.
+   * @returns Promise<PoiCategory> A promise that resolves to the created POI category.
    */
   createPoiCategory(poiCategoryForm: PoiCategoryForm): Promise<PoiCategory> {
     return this.apiBase.post<PoiCategory>({
@@ -522,7 +523,7 @@ export default class CartographyApi {
    * Deletes a POI category.
    *
    * @param {number} poiCategoryId - The ID of the POI category to delete.
-   * @return {void} A Promise that resolves when the deletion is successful.
+   * @returns void A Promise that resolves when the deletion is successful.
    */
   deletePoiCategory(poiCategoryId: number) {
     return this.apiBase.delete({