+++ test/domains/realtime.test.ts
@@ -3,28 +3,15 @@ import { UUID } from "../../src/types";
 import { getMockData, mockAxiosRequest } from "../utils/mockUtils";
 
 describe("SitumSDK.realtime", () => {
-  let situmSDK = null;
-  let mockRealtime = null;
-  let axiosMock = null;
-
-  beforeEach(() => {
-    situmSDK = new SitumSDK({ auth: getMockData("auth") });
-    mockRealtime = getMockData("realtimeMock1");
-  });
-
-  afterEach(() => {
-    situmSDK = null;
-    axiosMock.mockClear();
-    axiosMock.mockRestore();
-  });
-
   it("should retrieve realtime positions by organization id", async () => {
+    const situmSDK = new SitumSDK({ auth: getMockData("auth") });
+    const mockRealtime = getMockData("realtimeMock1");
     // Arrange
-    axiosMock = mockAxiosRequest([
-      { access_token: "fakeJWT" },
+    const axiosMock = mockAxiosRequest([
+      { access_token: getMockData("jwtMock") },
       getMockData("realtimeResponseMock1"),
     ]);
-    const organizationId: UUID = "1a5d0ff4-cd76-4e08-bcdc-e36e0182fd78";
+    const organizationId: UUID = "fakefakefa-fake-fake-fake-fakefakefake";
 
     // Execute
     const realtimePositions = await situmSDK.realtime.getPositions({
@@ -37,12 +24,16 @@ describe("SitumSDK.realtime", () => {
       `/api/v1/realtime/organization/${organizationId}`,
     );
     expect(realtimePositions).toEqual(mockRealtime);
+    axiosMock.mockClear();
+    axiosMock.mockRestore();
   });
 
   it("should retrieve realtime positions by building id", async () => {
     // Arrange
-    axiosMock = mockAxiosRequest([
-      { access_token: "fakeJWT" },
+    const situmSDK = new SitumSDK({ auth: getMockData("jwtMock") });
+    const mockRealtime = getMockData("realtimeMock1");
+    const axiosMock = mockAxiosRequest([
+      { access_token: getMockData("jwtMock") },
       getMockData("realtimeResponseMock1"),
     ]);
     const buildingId = 1111;
@@ -56,15 +47,20 @@ describe("SitumSDK.realtime", () => {
     const configuration = axiosMock.mock.calls[1][0];
     expect(configuration.url).toBe(`/api/v1/realtime/building/${buildingId}`);
     expect(realtimePositions).toEqual(mockRealtime);
+    axiosMock.mockClear();
+    axiosMock.mockRestore();
   });
 
   it("should retrieve realtime positions by organization id but without pass organization", async () => {
+    const situmSDK = new SitumSDK({ auth: getMockData("jwtMock") });
+    const mockRealtime = getMockData("realtimeMock1");
+
     // Arrange
-    axiosMock = mockAxiosRequest([
+    const axiosMock = mockAxiosRequest([
       { access_token: getMockData("jwtMock") },
       getMockData("realtimeResponseMock1"),
     ]);
-    const organizationId = "0a5d0ff4-cd76-4e08-bcdc-e36e0182fd78";
+    const organizationId = "fakefakefa-fake-fake-fake-fakefakefake";
 
     // Execute
     const realtimePositions = await situmSDK.realtime.getPositions();
@@ -75,5 +71,7 @@ describe("SitumSDK.realtime", () => {
       `/api/v1/realtime/organization/${organizationId}`,
     );
     expect(realtimePositions).toEqual(mockRealtime);
+    axiosMock.mockClear();
+    axiosMock.mockRestore();
   });
 });